refactor(hooks): files with hooks are moved into one folder
