+++ src/SmartComponents/PackageDetail/PackageDetail.test.js
@@ -16,11 +16,6 @@ jest.mock('react-redux', () => ({
     useSelector: jest.fn()
 }));
 
-jest.mock('../../Utilities/Hooks', () => ({
-    ...jest.requireActual('../../Utilities/Hooks'),
-    useFeatureFlag: jest.fn(() => true)
-}));
-
 jest.mock('../../SmartComponents/PackageSystems/PackageSystems', () => () => <div></div>
 );
